Added Tiles for awareness. Added Fireblast to main Proj.
Ive added white tiles round the edge of the map to help us to know where we are on the level when testing. Ive also tried implementing the fireblast mechanic in the main file. Ive changed the button click used to RMB as space is used to jump. I dont know if something in the original script is missing, as the test scene included by Daniel couldnt load the assets - Looks like an upload issue, but we can work that out when we next meet.